backend/requirements.txt
fastapi
uvicorn[standard]
tortoise-orm
asyncpg
python-jose[cryptography]
pydantic
pydantic-settings
python-dotenv
redis>=5
email-validator
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
